tests/wpt/web-platform-tests/css/css-color-adjust/meta/meta-color-scheme-single-value-in-shadow-tree.html

<!doctype html>
<title>CSS Color Adjust Test: meta color-scheme in shadow-tree should not apply</title>
<link rel="help" href="https://drafts.csswg.org/css-color-adjust-1/#color-scheme-meta">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/compute-root-color-scheme.js"></script>
<script>
  const host = document.createElement("div");
  host.id = "host";
  document.head.appendChild(host);
  const root = host.attachShadow({mode:"open"});
  const meta = document.createElement("meta");
  meta.setAttribute("name", "color-scheme");
  meta.setAttribute("content", "dark");
  root.appendChild(meta);

  assert_root_color_scheme("normal", "Meta color-scheme in shadow tree does not apply.");
</script>
