third_party/blink/web_tests/external/wpt/css/css-color-adjust/meta/meta-color-scheme-single-value-in-shadow-tree-expected.txt

This is a testharness.js-based test.
FAIL Meta color-scheme in shadow tree does not apply. assert_equals: expected (string) "normal" but got (undefined) undefined
Harness: the test ran to completion.
